Factor git stuff into module
