+++ lib/install.js
@@ -2,7 +2,8 @@ var path = require('path'),
     restler = require('restler'),
     shell = require('shelljs'),
     fs = require('fs'),
-    _ = require('underscore');
+    _ = require('underscore'),
+    Git = require('../lib/git');
 
 console.show = function () {
     console.log(arguments);
@@ -113,7 +114,7 @@ module.exports = {
             data.path = 'https://github.com/'+packageInfo.data.GitHubPath+'.git';
             data.dest = packagePath;
 
-            if (that._gitCloneRepo(data) !== 0) {
+            if (Git._gitCloneRepo(data) !== 0) {
                 console.show('Error: failed to clone from GitHub');
                 return fn({error : true});
             } else {
@@ -121,12 +122,6 @@ module.exports = {
                 return fn(null, packagePath); // send current path back up to trace dependencies
             }
         });
-    },
-
-    _gitCloneRepo : function (opts) {
-         var code = (shell.exec('git clone ' + opts.path + ' ' + opts.dest).code); 
-         console.show("GCR", code);
-         return code;
     },
 
     // get package info from registry
@@ -199,44 +194,5 @@ module.exports = {
                                     });
             }
         }
-    },
-    
-    // find branch most associated with sencha framework and version
-    _findBestBranch : function (opts) {
-        var senchaStuff = opts.senchaInfo,
-            branches =  this._gitGetAllBranches(opts.dest),
-            framework
-
-        console.show("FBB 1", senchaStuff, branches);
-    },
-
-    _gitGetAllBranches : function (opts) {
-        console.show("GGA1", opts);
-        var ran,
-            out;
-
-        shell.cd(opts.dest);
-        ran = shell.exec("git for-each-ref --format='%(refname:short)'");
-        console.log(ran.output.split('\n'));
-        out = ran.output.split('\n');
-        _.each(out, function (ref) {
-            ref.replace(/(\n|\r)+$/, '');
-        });
-        // parser(ran.output, function (err, data) {
-        //     if (err) console.log (err);
-        //     console.show("GGA2", data);
-        //     out = data;
-        // });
-        return out;
-    },
-
-    // checkout the required version of the package
-    _gitCheckoutBranch : function (opts) {
-        var git1,
-            git2;
-        shell.cd(opts.path);
-        git1 = shell.exec('git checkout -b ' + opts.branch);
-        git2 = shell.exec('git pull origin ' + opts.branch);
-        return ([git1, git2]);
     }
 };
